refactor(tools): fold literal edit primitives
